feat: refactoring validator to prepare for merging nodes
