+++ planning/autoware_trajectory_validator/include/autoware/trajectory_validator/filter_context.hpp
@@ -15,29 +15,6 @@
 #ifndef AUTOWARE__TRAJECTORY_VALIDATOR__FILTER_CONTEXT_HPP_
 #define AUTOWARE__TRAJECTORY_VALIDATOR__FILTER_CONTEXT_HPP_
 
-#include <autoware_perception_msgs/msg/predicted_objects.hpp>
-#include <autoware_perception_msgs/msg/traffic_light_group_array.hpp>
-#include <geometry_msgs/msg/accel_with_covariance_stamped.hpp>
-#include <geometry_msgs/msg/pose_stamped.hpp>
-#include <geometry_msgs/msg/twist_stamped.hpp>
-#include <nav_msgs/msg/odometry.hpp>
-
-#include <lanelet2_core/LaneletMap.h>
-
-#include <memory>
-
-namespace autoware::trajectory_validator
-{
-
-// Base context that all filters can access
-struct FilterContext
-{
-  nav_msgs::msg::Odometry::ConstSharedPtr odometry;
-  geometry_msgs::msg::AccelWithCovarianceStamped::ConstSharedPtr acceleration;
-  std::shared_ptr<lanelet::LaneletMap> lanelet_map;
-  autoware_perception_msgs::msg::PredictedObjects::ConstSharedPtr predicted_objects;
-  autoware_perception_msgs::msg::TrafficLightGroupArray::ConstSharedPtr traffic_light_signals;
-};
-}  // namespace autoware::trajectory_validator
+#include "autoware/trajectory_validator/evaluation_context.hpp"
 
 #endif  // AUTOWARE__TRAJECTORY_VALIDATOR__FILTER_CONTEXT_HPP_
